requirements.txt
Django==2.0
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.7.7
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
psycopg2==2.7.3